requirements.txt
# Core API
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1

# Database
supabase==2.0.0

# OpenAI (still used for embeddings)
openai==1.30.0

# Anthropic (Claude Haiku for chat/generation)
anthropic>=0.84.0

# Auth
python-jose[cryptography]==3.4.0
bcrypt==4.2.0

# HTTP
requests==2.32.0

# Utils
pydantic==2.7.0
chromadb==0.4.24
rapidfuzz
beautifulsoup4==4.12.3
